requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
opencv-python>=4.5.0
pillow>=8.0.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
torch>=1.10.0
transformers>=4.20.0
lifelines>=0.27.0
joblib>=1.1.0 